apps/core/pyproject.toml
[project]
name = "example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autoflake>=2.3.1",
    "black>=25.1.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flake8>=7.2.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "jinja2>=3.1.6",
    "mypy>=1.15.0",
    "openai>=1.75.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
namespace_packages = true
explicit_package_bases = true
exclude = ['venv', '.venv', '__pycache__', '.pytest_cache'] 

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
